+++ app/src/main/java/com/example/bookstory/DAO/relations/BookAuthorCrossRef/BookAuthorCrossRef.java
@@ -5,11 +5,11 @@ import androidx.room.Entity;
 
 @Entity(primaryKeys = {"bookId", "authorName"})
 public class BookAuthorCrossRef {
-    public int bookId;
+    public long bookId;
     @NonNull
     public String authorName;
 
-    public BookAuthorCrossRef(int bookId, @NonNull String authorName) {
+    public BookAuthorCrossRef(long bookId, @NonNull String authorName) {
         this.bookId = bookId;
         this.authorName = authorName;
     }
